Added preliminary online features, like having players joining into a game and starting the game.
Added two backend files and a frontend file to facilitate the players joining into a game and starting the game. Currently, the confirm move and the game settings are not working for the moment. But in conditions, the players should be seeing the same board with the same amount of balls. Give or take a couple of days.