Added moves feature to the online and win conditions
This feature allows the whole game to function properly. When the player confirms the move choices, the move is added to the redis backend to be pulled back around a second to the frontend, so the other player sees it. There are some issues to this commit. First, the win conditions do nothing other than to end the game. Second, there's a condition where the player's choice of looking at a previous move being overriden, depending on the player.